requirements.txt
boto3
pydub
google-cloud-texttospeech
numpy
librosa
soundfile
langdetect
